+++ src/guestbooks/domain/models/Guestbook.ts
@@ -25,4 +25,6 @@ export interface Guestbook {
   customQuestions: GuestbookCustomQuestion[]
   createTime: string
   dataverseId: number
+  usageCount?: number
+  responseCount?: number
 }